+++ models/tenant/tenant.go
@@ -134,7 +134,7 @@ func (u *User) Validate(isUpdate bool) error {
 	if !isUpdate && len(u.REV) != 0 {
 		return errors.New("Invalid Tenant User request: must not provide a revision value in a creation request")
 	}
-	if isUpdate && (len(u.REV) == 0 || u.CreatedTimestamp == 0) {
+	if isUpdate && (len(u.REV) == 0) {
 		return errors.New("Invalid Tenant User request: must provide a createdTimestamp and revision for an update")
 	}
 
@@ -182,7 +182,7 @@ func (d *Domain) Validate(isUpdate bool) error {
 	if !isUpdate && len(d.REV) != 0 {
 		return errors.New("Invalid Tenant Domain request: must not provide a revision value in a creation request")
 	}
-	if isUpdate && (len(d.REV) == 0 || d.CreatedTimestamp == 0) {
+	if isUpdate && (len(d.REV) == 0) {
 		return errors.New("Invalid Tenant Domain request: must provide a createdTimestamp and revision for an update")
 	}
 
@@ -219,7 +219,7 @@ func (prf *IngestionProfile) Validate(isUpdate bool) error {
 	if !isUpdate && len(prf.REV) != 0 {
 		return errors.New("Invalid Tenant Ingestion Profile request: must not provide a revision value in a creation request")
 	}
-	if isUpdate && (len(prf.REV) == 0 || prf.CreatedTimestamp == 0) {
+	if isUpdate && (len(prf.REV) == 0) {
 		return errors.New("Invalid Tenant Ingestion Profile request: must provide a createdTimestamp and revision for an update")
 	}
 
@@ -298,7 +298,7 @@ func (prf *ThresholdProfile) Validate(isUpdate bool) error {
 	if !isUpdate && len(prf.REV) != 0 {
 		return errors.New("Invalid Tenant Threshold Profile request: must not provide a revision value in a creation request")
 	}
-	if isUpdate && (len(prf.REV) == 0 || prf.CreatedTimestamp == 0) {
+	if isUpdate && (len(prf.REV) == 0) {
 		return errors.New("Invalid Tenant Threshold Profile request: must provide a createdTimestamp and revision for an update")
 	}
 
@@ -412,7 +412,7 @@ func (mo *MonitoredObject) Validate(isUpdate bool) error {
 
 	}
 
-	if isUpdate && (len(mo.REV) == 0 || mo.CreatedTimestamp == 0) {
+	if isUpdate && (len(mo.REV) == 0) {
 		return errors.New("Invalid Tenant Monitored object request: must provide a createdTimestamp and revision for an update")
 	}
 
@@ -487,7 +487,7 @@ func (meta *Metadata) Validate(isUpdate bool) error {
 	if !isUpdate && len(meta.REV) != 0 {
 		return errors.New("Invalid Tenant Metadata request: must not provide a revision value in a creation request")
 	}
-	if isUpdate && (len(meta.REV) == 0 || meta.CreatedTimestamp == 0) {
+	if isUpdate && (len(meta.REV) == 0) {
 		return errors.New("Invalid Tenant Metadata request: must provide a createdTimestamp and revision for an update")
 	}